Enhance SCOM's Linux log file monitoring capability
Use omsagent's Fluentd capability to enhance SCOM log monitoring.
Register SCOM as secondary workspace on the agent machine.
The Fluentd process of this workspace to be managed by SCOM.
New certificate is generated for securing the communication channel.
This certificate is used for authenticating agent to SCOM server.
Added a utility to regenerate the certificate.
5 filter plugins has been added that will generate events for SCOM.
An output plugin has been added to send the events to SCOM server.